Rename hugr/replace/{rewrite.rs -> replace.rs}
